modified mapbox's geocoder (open source)
The geocoder would take a user's location and use mapbox GL's flyTo() method to transition to this location.  This transition is too much 
to process in a mobile browser.  Modified this to use mapbox GL's jumpTo() method, specifying a zoom value of 7.  The hope with the zoom: 7 is that this will be zoomed in enough that there aren't as many zip tiles to process when the new location is drawn